Rename ControlPanel to SettingsPanel and remove VSegmentedControl
