+++ clients/macos/vellum-assistant/Features/MainWindow/SidePanelType.swift
@@ -1,7 +1,7 @@
 enum SidePanelType: Hashable, CaseIterable {
     case generated
     case agent
-    case control
+    case settings
     case directory
     case debug
     case doctor
